+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - bug fixes
     - server: fix resolving absolute file urls containing the remote-root path
+    - prevent editor zoom level from affecting drag events when the editor is disabled
 
 - widgets
     - the html property is no longer restricted / sanitized